Disable hover state for footer links in columns
Unfortunately there's a bug in Chrome which means links are mis-painted on hover when they're within CSS columns [1]. However, most service teams do not use the multi-column layout in the footer. We know GOV.UK do use it, but are already using a different implementation which does not use CSS columns. Therefore, disable the hover style for links within columns as think this feature is rarely used, if at all.

[1]: https://bugs.chromium.org/p/chromium/issues/detail?id=1190987