Add inverse link mixin and modifier class
Inverse links are a fairly common thing for service teams to need to implement, and we often see issues with the focus state being implemented incorrectly.

The upcoming changes to link styles will add even more nuance, so providing a reusable mixin and modifier class will save users time and help to make it more likely that link styles are implemented well.

We can also use this new mixin when we update the header component.