Tidy up redundant code in accordion
The CSS within the `hover: none` block is there to ‘undo’ the hover state for the `govuk-accordion__section-header` class on devices that don’t support hover (like touch screen devices) in order to prevent the hover state being applied on touch and getting ‘stuck’.

However, the hover state on `govuk-accordion__section-header` was removed in 9bfbc80 in favour of applying the hover state directly to `govuk-accordion__section-button`, so there’s nothing to ‘undo’ any more and this block can be removed.

This is the only place where `$govuk-accordion-border-width` gets used, so we can remove that as well.