Update accordion to use new link styles
Simplify the code by removing the `accordion-link-colour` abstractions.

Move the colour definition to the `section-button` rather than inheriting it from the `section-header`, which I believe is leftover from when the `section-header` was the element that was ‘hovered’, and thus we needed to inherit the colour depending on hover state.

Apply link styles to the `section-button` on hover. There’s some duplication here, but we know the accordion component is due for some major updates soon so avoid worrying about it for now.