Work around link hover bug in Safari
There’s a bug in Safari that means that changes to text-decoration-thickness on hover are not applied above certain font sizes, unless there is also another change – like a change to the text colour. [1]

Add a text colour change to the muted link style by making it black (text-coloured) on hover.

Force a text colour change for text and inverse links (which currently don’t change colour) by setting the alpha channel to 99% on hover. This should generally be imperceivable, but is enough to force Safari to paint the hover style correctly.

[1]: https://bugs.webkit.org/show_bug.cgi?id=224483