Set link underline thicknesses in both px and rem
When GOV.UK Frontend is used in compatibility mode alongside GOV.UK Elements, Elements sets the root font size to 10px.

In rem, the thickness of underlines in the hover state with a 10px root font size is 1.875px (10/16ths of 3) rather than 3px and so the hover state does not display correctly.

Include both px and rem, as this means that in situations where the root font size is less than 16px, the value in px will 'win' and so the underlines are displayed correctly.