Fix link text for anchor links to GDS Way pages
Commit a4c057 updated the URL for links to style guides that had been
moved to the GDS Way. In doing so the links now violate the rule that
anchor links should not be used without explaining that the link goes
into a page [1].

This commit updates the link text so that it is clear that the links
take you to a place within a page.

[1]: https://docs.google.com/document/d/1XY8RG2uQjQNo7PFhtjvqIcHqGqSgGDRuQGkpOFpc6iM/preview

Co-authored-by: Eoin Shaughnessy <eoin.shaughnessy@digital.cabinet-office.gov.uk>