Fix checkbox display in IE8
We use the `before` and `after` pseudo-elements on the label for checkboxes to provide stylised checkboxes in modern browsers.

Internet Explorer 8 does not support some of the features we use to do this, so we ‘fall back’ to the native HTML inputs instead.

In 09ef79d2e4ed5ea824d1f09526f902a4c8edc643 we updated the before and after pseudo-elements in the checkboxes sass to use a single colon, which was newly enforced as part of a switch to stylelint.

IE8 only supports the single-colon syntax [1], and it looks like we were actually relying on the double-colon syntax _not_ being interpreted by IE8 in order to prevent them being displayed in that browser.

As a result, changing to a single-colon means that IE8 now interprets the pseudo-selectors, and as a result ends up with a malformed checkmark visible next to the native HTML input.

Wrap the before and after pseduo-elements with the @govuk-not-ie8 mixin to exclude these rules from being included in the IE8 specific stylesheet.

[1]: https://caniuse.com/css-gencontent