Fix GOV.UK logo disappearing in Windows HCM
In 6f90027 I updated the logo in the header to use `forced-color-adjust: none;` in order to fix an issue where the readability backplate obscured the underline on the link.

However, I neglected to then handle the color adjustment correctly – the logo is now always white in high contrast mode, regardless of the user’s colour preferences. This means that the logo can be invisible or hard to perceive when the user’s colour scheme has a lighter background.

To fix this, set the colour of the logo to the system color `linktext` when forced colors are active, which means it will appear in the user’s preferred colour for unvisited links, matching the appearance of the other links in the header.