Fix diff changes to dist GitHub Action
Description of the issue
------------------------

When raising a release PR, we have a GitHub Action that is meant to add
a diff of the changes to dist (what a mouthful!) as a comment on the PR.

This failed when we were trying to release v3.12.0 with `Error: Argument
list too long`.

See https://github.com/alphagov/govuk-frontend/pull/2224/checks?check_run_id=2575562845

Fix
---

Following a suggestion from another person who had the same issue [[1]],
this commit saves the diff text to a file in the GITHUB_WORKSPACE folder
instead. The GitHub Actions states actions can modify this directory and
have subsequent actions access those changes [[2]].

This also has the advantage of removing the need to percent-encode the
diff text, so the script to generate the diff can now be simpler.

[1]: https://github.community/t/maximum-length-for-the-comment-body-in-issues-and-pr/148867/5
[2]: https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#file-systems