Increase `git diff` similarity index threshold
Changes the similarity index threshold for detecting renames to 5% (see
man page for git diff for more details on the `--find-renames` option).

For release 3.12.0 the similarity index for the minified CSS was lower
than 10%, so the diff output presented the changes to the CSS as a
deletion and a creation, which isn't very useful. Using `-M05` to set
the threshold made output more readable, so we should keep that change.