Don’t remove focus outline from disabled link buttons
SortSite flags this as a potential issue, referencing WCAG 2.0 F78 [1]:

> The CSS outline or border style on this element makes it difficult or impossible to see the dotted link focus outline.

Adding the `--disabled` modifier to link buttons (`<a>` elements styled as buttons) does not remove them from the tab index or make them non-interactive. From testing, they are still focusable and retain the rest of the focus state, so avoid trying to remove the outline.

This change should not effect buttons, as disabled buttons are not focusable and so the focus pseudo-selector should never apply to them.

[1]: https://www.w3.org/TR/WCAG20-TECHS/F78.html