Avoid rounding errors in 1/3 and 2/3 grid widths
Use 6 decimal places for grid percentages to avoid fractional pixel widths caused by rounding errors. This is mostly a neatness thing, but _may_ avoid instances of unneccesary sub-pixel rendering caused by the fractional pixels.

## Chrome

The calculated width of a two-third grid column changes from 659.98px to 660px, and a one-third grid column from 329.98px to 330px.

## Firefox

The calculated width of a two-third grid column changes from 659.983px to 660px, and a one-third grid column from 329.983px to 330px

## Safari

There is no reported change in the calculated width, but it does seem to fix an issue where the kerning between the first and second letter of each line in a one-third column is slightly looser than it should be

## Internet Explorer

There is no change – a two-third grid column is 629.93px and a one-third column is 299.96px. I believe because IE11 truncates to 2 decimal places [1] so adding more decimal places to the CSS makes no difference.

For each browser, the width was found by inspecting the element using dev tools with the page at full width.

[1]: https://cruft.io/posts/percentage-calculations-in-ie/