Add user adjusted colours
When a user adjusts their colour scheme to meet their needs (a11y) the original styles make the button unreadable. The background, the icons and the text all become the same colour once interacted with.

This fix allows an adjustment for or user preferred colours (such as high contract, low contrast and other colour adjustments)

- On hover add underline to toggle text to add extra affordance as icon colour changes on pseudo element are not respected.
- On focus add transparency to background so content is visible
- Update comments

This addition has been based on code within govuk-frontend

[1] https://github.com/alphagov/govuk-frontend/blob/main/package/govuk/components/header/_index.scss#L45

and then extended for wider coverage based on:

[2] https://github.com/material-components/material-components-web/blob/f147a2271bba2b4f1ae4df403baf86bac974b120/packages/mdc-dom/_dom.scss