Change summary element
The Design has been updated to include the sub-text or "summary" of the Accordion section Content to be directly underneath or after the Headline.

(@nnagewad) "We’ve decided to move the summary in order to better visually illustrate the additional context of accordion heading by placing it directly beneath the heading as opposed to placing it directly beneath the reveal/expand button"[1]

Once JS is loaded the summary element is converted from div to a span.

As there is a nested element within a button <div> is not semantically correct or ideal in this context.  This has been altered to span element and all relevant properties and Content as also transferred.

[1] https://lawsofux.com/law-of-proximity/