Update spacing while focused
While elements are focused on mobile - the individual focuses states have a tendency to overlap.  

This is partially because the "focus state Design" is slightly shorter than the actual focus styles that is in use[1]

Add some spacing between various elements on mobile. However this also enlarges to the total overall height of the element slightly.

[1]https://www.figma.com/file/3yTE9s4EOkvjcqRnqaIzgG/Accordion-Redesign?node-id=343%3A726