Fix test isolation of accordion component tests
The unit tests for the accordion component can interfere with each
other, because within a test suite Puppeteer browser instances are
shared, and the accordion component has state stored in the browser
session storage. This is especially a problem for any tests after
`it('should maintain the expanded state after a page refresh')`, which
leaves all the accordion sections opened.

This commit adds a simple teardown function to clear the session storage
after each test. It just so happens that our current tests behave the
same and still pass with the teardown function, but we should fix this
problem now, to prevent any future tests from being affected by the
state of our current tests, such as for the accordion design refresh in
PR #2257.