Add failing tests for IDs with special characters
When checkboxes with conditional reveals are initialised, we check that the element targeted by aria-controls exists:

```
if (!controls || !$module.querySelector('#' + controls)) {
   return
 }
```

Because the string passed to `querySelector` is evaluated as a CSS selector, this can fail if the ID contains characters that have a special meaning in CSS, such as `.` or `[]` – the ID would need to be escaped for it to be evaluated correctly.

Add a failing test to cover this case.

The tests fail with:

```
Checkboxes with conditional reveals › when JavaScript is available › does not error when ID of revealed content contains special characters

    DOMException: Failed to execute 'querySelector' on 'Element': '#conditional-user.profile[contact-prefs]-2' is not a valid selector.
```

```
Radios with conditional reveals › when JavaScript is available › does not error when ID of revealed content contains special characters

    DOMException: Failed to execute 'querySelector' on 'Element': '#conditional-user.profile[contact-prefs]-2' is not a valid selector.
```