Change how focus state is set
The CSS adjustment removed in #132c12f had the favourable side effect of limiting the width of the focus state for the three text items (heading, summary, show/hide), stopping the focus state from expanding beyond the text to the full width of the component.

To limit the width of the focus state without the NVDA adjustment, this commit:
- adds focus containers in the markup inside the three items using JavaScript
- makes the focus containers inline items
- moves the focus styles to those classes; the inline style sets the width of the focus state to the width of the text
- renames `.govuk-accordion__section-heading-focus-wrapper` to `.govuk-accordion__section-heading-text` since the class no longer has the focus styles

Change the text containers to set bottom margin instead of top margin as this:
- resolves an issue with the summary line margin: since the summary line container is now a block level one, it renders the `.govuk-body` bottom margin; moving the margin to the bottom of the container makes the container margin collapse with the `.govuk-body` bottom margin
- is more consistent with how rest of GOV.UK Frontend sets margins

As the last text container inside the butotn now has a bottom margin, we also don’t need set the bottom margin on the button anymore.

Also remove a now redundant rule that would add an underline to improve affordance for text inside pseudo elements since we no longer use pseudo elements for text content inside the button.