Revert removing the summary ID from accordion template
We originally intended to remove the ID from the template markup as it's no longer required by `aria-describedby` in the accordion section header.

Revert that change since we've decided not to ask teams to make this breaking change yet. We want to wait to see if we receive any
feedback from teams regarding the issue flagged in https://design-system.service.gov.uk/components/accordion/#known-issues-and-gaps
(refer to https://deploy-preview-1961--govuk-design-system-preview.netlify.app/components/accordion/#known-issues-and-gaps
until we publish the v4 guidance in the Design System).