Use document.getElementById to get count message
Because the string passed to `querySelector` is evaluated as a CSS selector, this could potentially fail if the ID contains characters that have a special meaning in CSS.

Because we're using an attribute (`[id="foo"]`) selector, and wrapping the ID in double quotes, the current approach does seem more robust than using an ID selector (`#foo`) as we have been doing in other components.

However, to be consistent with other components and to avoid other potential issues, switch to `document.getElementById` instead, so the ID is no longer evaluated as a selector.

The downside is that we can't constrain the scope to the $module. However, we don't think this should cause any issues as the ID _should_ be unique within the document.