Replace use of grid with flexbox
Based on [[1]], we can't use grid autoplacement in any version of
Internet Explorer, so we have to go back to flexbox.  We need to use
border-box sizing and padding to make the gutter work the way we want,
instead of using margins.

[1]: https://css-tricks.com/css-grid-in-ie-faking-an-auto-placement-grid-with-gaps/

This also lets us support the (old) GOV.UK footer layout, where there were
two rows, one with a single 2 column section that took 2/3rds of the
space (followed by empty space), and a second with a 2
column section and a 1 column section. Having `flex-grow` would mean
that the 2 column section in the first row would take up all the space.

This is also closer to what the grid layout does out of the box, so
that's also good.

Note: we're using `width` instead of `flex-basis` to support IE11

See [flexbug #7] for more details on why this is
needed.

flexbug #7: https://github.com/philipwalton/flexbugs#flexbug-7