Use single column footer on mobile viewports
We want the footer to switch from a multi-section
multi-column affair, to a one-section-per-row single-column-per-section
layout on mobile.

This commit adds a media query so that the switch
happens at the tablet breakpoint.

However, there's lots of space in between the mobile layout and the
desktop layout for all sorts of responsive designs, which this commit
eschews; so we might want to revisit this.