Bugfix: stop arrow from receiving click events
The arrow was being positioned over the input using z-index, which meant it received click events which stopped the suggestions being fired.

One solution to avoid this is to position it underneath the input (using z-index) but to make the input background transparent. However this has a whole host of other issues.

A simpler alternative for now is to stop click events from firing by setting `pointer-events: none;`. This doesn't work in IE <11, so there might need to be some workarounds for that browser - eg adding the click handler to that element.