Remove display block style on hint component
In v3.8.0 we added the ability for block level elements to be
rendered inside hints by switching from using a `span` to a `div`.

However, we kept the `display: block` styling (which would now be
redundant as this is the default for divs). This meant we weren't
forcing users to update their HTML and we could avoid a breaking change.

We now want to tidy up that redundant CSS as part of v4.0.0.