Update pre-release instructions to use npm ci
Ask users to run `npm ci` instead of `npm install`. Whilst `npm ci` will take a bit longer to run,
it will not force any dependencies to update - at this point we just want to ensure that we are using
the already installed dependencies.

Also make the wording of step 1 more ambigious as the current command will create a new branch and this
might not be appropriate.