Update legend/label/heading examples to have unique names and ids
Every example in the legend/label/heading examples has the same `name`
and `idPrefix`. Having the same `name` on every example meant that the
checkboxes and radio buttons looked like they were part of one question.
For example, when testing with Voiceover and Safari, you would hear something
like:

"Yes, radio button, 1 of 60" for a simple yes/no question.

The `idPrefix` is used to prefix the id's for the hints. If not present, the `name`
is used instead. Because the `idPrefix` values were also identical, this led to hints being
associated and read out for the wrong questions because there were multiple hints with
the same `id`.

This commit updates each `name` to give it a unique value. It also removes
the `idPrefix`, which means the `id` will be generated from the `name`. This is ok
in this case as the names are now all unique.