Replace text input blank `id` with unique value
The `id` for text input examples were being set to a blank string.
This was leading to text inputs having an `id` (and therefore an `aria-describedby`)
starting with a dash, i.e: `id="-hint"`. Each text input also had the same `id`.

This commit replaces the blank `id` and uses the same value as the `name` so each
text input has a unique `id`.