Make list margins above and below equal
The previous implementation had the margin above the list set to 40px on desktop
and 25px on mobile, and the margin below the list set to 30px at all breakpoints.

This commit makes the margins equal - it reduces the top margin to 30px as it was
felt that the 40px made the list too far away from the heading.