Don't pre-select radio buttons
The radio button examples had the 'No' radio pre-selected, but we don't
recommend pre-selecting in our guidance and none of the other examples
on this page have any preselected/pre-populated inputs.

The `checked` attribute was already present in the code, but wasn't working
properly because of the duplicate names/ids. Now that we fixed that, it had
the unintentional effect of making the `checked` code work.

This commit removes the `checked` attribute from the radio button examples.