Make list margins above and below equal
Each list in the footer section as a bottom margin of 30px, and a top
margin (which is actually set by a bottom margin on the heading) of
40px on desktop and 30px on mobile. This means that on desktop the
spacing above and below the list is unequal.

Having spoken to Chris, we've decided to make this spacing and drop the
heading margin down to 30px on all breakpoints as it felt too large.