Rename and update footer YAML examples
Update the footer YAML examples to use the new `width` macro option

Update the full GOV.UK yaml example to reflect the current layout
and use the new `width` macro option.

This commit also renames the example. When the footer example
has 'GOV.UK' in the name, it gets
flagged by Chrome as a possible spam/phishing site which
means you can't open that example in a new window.

This commit works around that issue by renaming the example -
it seems to be the word "GOV.UK" which triggers it