Fix tests for jest v27
jest@27.0.0 included a breaking change that calling `done()` and
returning a Promise cannot be combined [[1]]. This commit fixes the
instances where we do this.

Note that in `src/govuk/all.test.js`, we're not (I think) actually using
any async code at all (the puppeteer docs don't show examples of using
`page.on` with await [[2]]). Thanks @36degrees for spotting this.

[1]: https://jestjs.io/blog/2021/05/25/jest-27#features-coming-with-breaking-changes
[2]: https://pptr.dev/#?product=Puppeteer&version=v12.0.1&show=api-class-page