Remove gulp-stylelint for direct stylelint integration
While looking at what would be needed for govuk-frontend to adopt a
stylelint 14 compatible release of styleint-config-gds [1] I discovered
that a blocker for that upgrade is the use of gulp-stylelint [2] which
only officially supports Stylelint up to version 13 and doesn't seem to
be actively iterated.

There didn't seem to be much that was lost by removing gulp-stylelint,
it actually made SCSS linting consistent with the approach taken for JS.
As the command to run stylelint is longer than standard I added
additional tasks to the NPM file that are consistent for both JS and
SCSS linting.

I've used quotes around the paths as it seems they were being expanded
by the OS rather than by stylelint and using quotes forces stylelint to
do so.

[1]: https://github.com/alphagov/stylelint-config-gds/pull/22
[2]: https://github.com/olegskl/gulp-stylelint/issues/132